Load common JS dependencies via LoadJS
The previous jQuery based loading of external scripts causes problems
when the scripts themselves tried to access additional resources. If an
external script A tried to load B, B was resolved with respect to the
script file that loaded A, not with respect to A itself. This led to
failed transitive loads. LoadJS takes proper care of this by creating
new <script src="A"> tags instead of retrieving the script text and
executing it in the currrent context.