Refactor inline script tags into ES6 modules
Moving most of the code into ES6 modules is a first attempt to get rid
of the spaghetti code structure in content-slider.html. The individual
modules still need to be refactored to rely less on global variables.

The refactored code also doesn't depend on jQuery anymore which was only
used for very few things anyway.