Fall back to the sample config only if parsing of given the config fails
If a configuration could be retrieved but contains errors, it should be
accepted as is such that the parse error in the given config can be
fixed. The fallback config will now only be used if the given config
file can not be retrieved e.g. due to a network error or because the
specified files just doesn't exist.