Fix flickering during inital page load
The page is faded in after a given amount of time. However, the page was
initially visible until a certain CSS class has been attached via
JavaScript such that there are a couple of ms where the page is fully
visible and this often leads to a kind-of disturbing flickering effect.
This change attaches the needed CSS class from within the HTML, but
initially pauses the animation via the animation-play-state CSS
property. The fade-in animation can now be started later via JavaScript
without any kind of flickering being present.