Reload page after fadeOut animation ended
Instead of relying on a timeout to reload the page after it has been
faded out, we now wait for the actual animationend event of the fadeOut
animation. This also means we don't need to adjust the timeout length
when the duration or delay of the fadeOut animation are adjusted.