Add option to auto-reload on JavaScript errors
The class ErrorReloader can be configured to reload the page whenever an
error event is emitted on the window (e.g. unhandled exceptions).
The reload will happen after a delay configured via the config option
reloadOnErrorDelay (seconds). A negative number can be set to disable
the auto reloading, e.g. for debugging purposes.