Reduce built-in keywords in anniversaries to %%ago(SPEC)%%
The only necessary keyword for inserting timespans is now %%ago(SPEC)%%,
where SPEC is a subset of the keywords DEFAULT, ALL, MILLENNIA,
CENTURIES, DECADES, YEARS, MONTHS, WEEKS, DAYS, HOURS, MINUTES, SECONDS
and MILLISECONDS join by the | character.

In addition, labels for each part of the timespan can be speficied in
the anniversary JSON file. Setting labels to the empty string can be
utilized to insert just the respective number.