Add component for displaying announcements
The AnnouncementManager will display announcement configured in a separate file after a configuration idle time of the station. It utilizes the content-slider-announcer component for the actual display of the announcemnts.

The JSON/YAML schema for the announcement file format is stored in schema/messagesSchema.yaml.

Additional Whenzel-filterable settings for content-slider-announcer can be configured in a separate config file.